Make type variants made by shapes consistent with schema inheriance rules
Currently, overriding any link or property in a type variant produced by
a shape with a type that is incompatible with that of the base type is
allowed.  This is inconsistent with the normal schema inheritance rules,
and could lead to surprising behavior if the client receives instances
of a type with property types that it doesn't expect.

The only scenario I can think of where the current behavior is useful, is
schema emulation for backwards compatibility, and that use case should
be covered by the future introduction of proper computed types.