Implement rudimentary support for type intersection
While working on tightening the restrictions around derivation of type
variants, I realized that our current formulation of the `[IS <type>]`
operator is mostly wrong, at least from the standpoint of the type
system.

Currently, the result type of `<expr>[IS Type]` is always `Type`, and,
this is only OK if `Type` is a subtype of `TYPEOF <expr>`.  In other
cases `[IS ...]` strips useful type information unnecessarily.

This stems from viewing `[IS ...]` as a "type filter", whereas in
reality it is a *type intersection* operator.  It returns an
intersection of its argument set with the set of all instances of
`Type`, so the result type is not just `Type`, it's an *intersection
type* `TYPEOF <expr> & Type`.  Intersection types are described in
detail in #219, but the most important detail is that an intersection
type contains a union of all properties and links of its constituent
types, so, if we had a schema like:

   type Named {
       property name -> str;
   }
   type Text {
       property text -> str;
   }

Then, `Named[IS Text]` would be a type that has *both* the `name` and
the `text` properties.  Similarly, if the argument is a union type, the
intersection operator may *narrow* the union, and in any case, the
intersection type is always a subclass of its consituent types -- a
property that is vital for polymorphic expressions in shapes.