Fix a number of issues in handling of link unions
Like in the previous commit, the main issue is incorrect handling of
unions of links that have overlapping descendants.  This is fixed by
applying the same technique as for type unions.

The second issue is caused by the fact that we currently don't allocate
tables for purely-inherited links.  This was originally done as an
optimization, but it is actually exceedingly hard to handle cases when a
purely-inherited link becomes local and obtains a table.  The data
becomes split between two tables, which are also linked by inheritance,
which means that it is possible for a query to produce duplicate
results.

Finally, we are forgetting to actually include an intersection
of link properties in the union link object.