Set commutator and negator on operator aliases explicitly
The majority of operators are defined as aliases for the corresponding
SQL operator via the `USING SQL OPERATOR` clause.  Currently, this
creates almost a complete copy of the base operator, possibly, including
its commutator and negator.  This is problematic for two reasons:

1) Postgres requires a user to be the owner of an operator used as a
   commutator or a negator (or a superuser), which might not be true,
   and

2) Blind copy of commutator and negator might be incorrect if the actual
   operators are not pure aliases.

Fix this by specifying the negator and commutator explicitly on all
operators that require them.