Use immutables.Map instead of a frozenset to store schema crossrefs
The memory profiler added in #1032 has quickly pointed to
`Schema._update_refs_to()` as being the largest memory consumer during
query compilation.  This is due to the use of `frozenset` to store the
ids of referring objects.  Replacing it with `immutables.Map` results in
a significant reduction of memory use.

BEFORE:

    compile_edgeql_script_webapp_get_movie.memory:
        2900.13KiB / 6610 blocks

AFTER:

    compile_edgeql_script_webapp_get_movie.memory:
        855.99KiB / 7227 blocks (*70.48% less*)