Add an option to inject an implicit limit into read queries
The client can now, via a special header to the Parse, request
the server to inject a constatnt limit into all read queries.
This is useful for contexts where the precise cardinality of
output is less important, and injecting limits ensures that
large linked sets in shapes don't bring everything down.

This requires edgedb/edgedb-python#67.

Fixes: #846.