edgeql: Forbid redefinition of read-only flag.
The read-only flag presents possible problems of interpretation if it's
allowed to be redefined in inheritance hierarchy. Specifically, consider
the following:

    type A {
        property foo -> str {
            readonly := True
        }
    }

    type B {
        property foo -> str {
            readonly := False;  # not omitted for clarity
        }
    }

    type C extending A, B;

If this is legal, then regardless of how we interpret the correct way to
inherit `readonly` in `C`, we will have unintuitive behavior of objects of
type `C`. It's not obvious to the end-user whether the property `foo` of
C should be more like A or like B.

The solution is to forbid mixing `readonly` flags much like it's
forbidden to mix or redefine cardinality via inheritance.

Issue: #1047