edgeql: Make it invalid to `DROP REQUIRED` if it wasn't defined.
`DROP REQUIRED` doesn't make sense when the required qualifier is not
set or is inherited. It can also lead to unintuitive results if it's
allowed since `DROP REQUIRED` cannot actually override an inherited `SET
REQUIRED`.

The solution is to only allow `DROP REQUIRED` in the contexts of a link
or property on which it was explicitly set.