Untangle the CREATE DDL path
Currently, the order of processing of `CreateObject` subtrees is rather
awkward.  The generic implementation simply iterates over referenced
object subcommands, and various schema classes tack on their own hacks
to implement inheritance propagation etc.  On top of that, handling of
anonymous derived objects, like unions and collection types is hacky and
buggy as well.  To fix this, make the generic `CreateObject.apply` truly
generic, which makes it behave like `AlterObject.apply` already does.
This necessiated the addition of a concept of "prerequisites" for a
command, i.e. subcommands that must run before the object itself is
created.