docs: Update the documentation for SDL.
Update the SDL overview documentaiton to indicate that the SDL describes
the schema as a whole as opposed to a single module.

Add a section to describe SDL module blocks.

Update migration docs and remove DDL migrations as we have tested the SDL
migrations only.

Make doc tests aware that SDL may or may not contain a module block and
thus require slightly different handling of the syntax test.

Fixes: #974.