Generalize preservation of original text in expression schema fields
Currently, the expression text used in the original object definition is
handled by a bunch of duplicating hardcoded logic in the corresponding
object implementation.  This generalizes the mechanism and also makes
the "orig_expr" fields ephemeral, since the place for the storate of the
original text is in the main Expression structure.  The "orig_" fields
now essentially serve as markers that the corresponding expression field
should preserve the original text.  With a bit of further development,
it might be possible to elide the definition of these shadow fields
altogether and enable this behavior on all expression fields.