Change the syntax of target type intersection in shapes
Currently, the syntax to specify the target type intersection in a shape
element expression is this:

    SELECT User {
        favorites: Book {  # Book is a type expression
            title
        }
    }

Despite being visually similar to the top shape, `Book` is not at all an
expression, it's a type expression equivalent to the inside of an
`[IS ...]` operator.  The confusion is further exacerbated by the fact
that changing the link from a traversal to a computable changes the
meaning of `Book` to be an expression:

    SELECT User {
        favorites:= Book {  # Book is an expression
            title
        }
    }

To avoid this confusion, change the target type intersection syntax in
shape navigation to use the `[IS ...]` syntax:

    SELECT User {
        favorites[IS Book]: {
            title
        }
    }

This has two nice properties:

a) We use `[IS ...]` consistently to express type intersection in all
   scenarios;

b) The shape construct becomes an almost perfect _folding_ of path syntax.
   Compare:

    AbstractUser[IS User].favorites[IS Book].title

and

    AbstractUser {
        [IS User].favorites[IS Book]: {
            title
        }
    }

The change is purely syntactic, since this is already represented as a
type intersection at the AST level.