Switch to C.utf8 locale by default
We currently use the bare "C" locale, which makes character handling
outside of the ASCII range not work as expected.  Switch to "C.utf8"
(but keep bare "C" for collation).  This also spells locale and encoding
explicitly for newly created databases.  Not all systems support C.UTF-8,
so fallback to en_US.UTF-8 in those cases.

Fixes: #1105.