graphql: Add bigint tests and fix some translation bugs.
Currently we don't distinguish between limited and unlimited
precision/size numbers in GraphQL. Which means that by default it
makes sense to treat `Int` as `bigint` and `Float` as `decimal`
(especially for variables). This preserves the precision when it's
needed and can be downcast to the specific type the expression
requires.

Since `int64` is implicitly castable into `bigint` it's possible to
treat small `Int` literals as `int64` and large ones as `bigint`.

Stop mapping `bigint` onto `Float` and use `Int` instead.

This technically breaks the sprict spec for `Int` and `Float`.
Introducing custom types would be a separate PR.

Issue #1138