tests: Adjust some test to check `decimal` exact value.
A few tests that return `decimal` values now explicitly check the "json"
and "binary" results. The "binary" protocol is supposed to preserve type
information and actually produce a `decimal.Decimal` in Python. However,
fetching data as "json" can only produce a `float`.

Update the docs for the datatypes `bigint` and `decimal` with a note
about potential problems with JSON conversion.

Fixes: #1151.