Fix erroneous scope construction for paths with computable links
When handling paths like `Tree.parent`, where `parent` is defined by an
expression, make sure that the scope tree is constructed such that
the `Tree` symbol is not visible from the computable expression (unless
it has been pulled into an outer scope).  In other words, make sure that
the `Tree.parent` expression is not treated as if it was part of
`(Tree, Tree.parent)` tuple.

Specifically, this means that instead of the following incorrect scope:

    (test::Tree).>parent[IS test::Tree]: {
        (test::Tree)
        "FENCE": {
            <expr-scope-for-"parent">
        }
    }

this is generated instead:

    (test::Tree).>parent[IS test::Tree]: {
        "BRANCH: {
            (test::Tree)
	},
        "FENCE": {
            <expr-scope-for-"parent">
        }
    }

Fixes: #1098