Fix several issues in handling of type intersections
The current handling of type intersections fails to account for the case
where a union is being narrowed by a *subtype* of one of the union types
(as opposed to a *supertype*, which is handled correctly).

Another issue is that the column routing code currently fails to account
for intersection relvars and just picks the first relation out of a set
of intersection type relations.  Fortunately, the "inherited" pointers
in compound types retain the information about the original defining
type, so it's relatively easy to pick the correct relvar by comparing
its typevar to the typevar of the pointer source.