Remove reliance on Postgres table inheritance
EdgeDB currently relies on PostgreSQL table inheritance to support the
generic query semantics as well as implement certain aspects of schema
information propagation.  While it made sense at the point of initial
implementation, nowadays the query compiler and the schema management
implementation in EdgeDB are significantly more mature, to the point
where Postgres' inheritance semantics actually get in the way.  Another
point against using table inheritance is that it is incompatible with
native partitioning, and we want to be able to expose that at some
point.

The removal of reliance on Postgres table inheritance requies emulating
its behavior in queries and in DDL.  For read queries we emulate
inheritance by generating SQL views which contain a union over the
target type table and the tables of its descendants.  Union views are
not updatable, so for DML we have to emulate by generating multiple
SQL DML statements for each concrete table.  The DDL side of things is
mostly untouched, and boils down to the removal of inhibition of the
inherited object reflection.