Allow tuple arrays in schema definitions
Currently, `array<tuple<...>>` is not allowed in schema definitions
(although expressions of such types are allowed in queries).
Originally, this restriction was placed because Postgres has
non-trivial limitations with respect to composite type casts.
Specifically, it is impossible to cast `record[]` to any composite type
array, and it is also impossible to cast a composite type into another
composite type even if they are compatible element-wise.

Fortunately, the casting infrastructure in the compiler is now advanced
enough to side-step the aforementioned issues and remove the tuple array
restriction.