Fix merging of link targets when subsequent base is contravariant
The current implementation of `merge_targets` only allows covariant
merging, i.e. if the subsequent target is a subtype of a current target.
This is correct when computing an object type subtype, but not when
merging multiple parents into a subtype.  In the latter scenario, the
covariance and contravariance of specific link targets in bases are
equivalent.

While at it, clean up the check code and add a few tests.