Use postgres/notify to invalidate query cache on DDL
This makes it possible to use different EdgeDB servers connected
to the same Postgres cluster.  With this change, when a DDL command
is committed, a `NOTIFY` command is issued which all EdgeDB
connections listen to.

This commit also introduces IDLE/ACTIVE states for pgcon. In IDLE state
pgcon consumes async messages such as notifications as they come.