graphql: Change the auto-generated type name template.
When object types are reflected into GraphQL, we create an interface
with the original type name and an underlying type implementing that
interface with a name that has "_Type" appended to it. The idea is that
since the convention in EdgeDB is to have camel-case type names adding
an "_" will result in fewer cases of type name clashes.

Fixes: #1175.