Handle tuple indirection as a regular path
Currently, tuple indirections are handled as expression paths (via the
`TupleIndirection` node), but this is redudant, since the `path_id` of
the indirection already contains all the information needed to process
such expressions.  It is also a source of bugs, as evidenced in #1255.
The solution is quite simple: drop `TupleIndirection` expression node
and treat tuple indirections as special path steps, akin to type
intersection steps.

Fixes: #1255.