Include psedo types in the schema
Currently, pseudo types (`anytype` and `anytuple`) exist as singleton
instances outside of the schema object.  The reasons are purely
historical and do not hold water anymore, since the schema can now
hold unqualified objects of any type.  On the other hand, the current
arrangement is inconvenient and requires hacks to work around.  Rectify
that by registering the pseudo types when the stdlib is populated.