Reimplement schema object shell mechanism
There are cases when schema objects cannot be referred to directly from
a schema delta: forward references and deltas resulting from schema
comparison.  In these cases we replace schema object references with
isntances of `ObjectRef` (or its numerous subclasses) that are a
schema-independent way of describing a schema object in *some* schema.
For most objects it's just a name.

The current implementation of `ObjectRef` is problematic because it
tries to masquerade as a regular schema `Object` (mostly for historical
typing reasons), but since it's not a real object, the schema code is
peppered with checks in places where `ObjectRef` is not expected, and,
since it masquerades as an `Object` the typing doesn't help much (or,
worse, gets in the way).

Luckily, with the current state of things, the use cases for object
shells are much more limited, and the above mentioned masquerading is
not necessary.  This commit adds a new implementation of `ObjectShell`
(and `TypeShell`, etc.) that does the necessary minimum in a much
cleaner way.

The old implementation is left in place to simplify the diff and will be
excised cleanly in a subsequent commit.