Remove obsolete guts of ComparisonContext
`ComparisonContext` is currently implemented as a stack of context
levels, similar to the context pattern elsewhere.  However, the
comparison context has been completely unused for quite some time.
Remove the stack guts and replace with a simple context object, since we
may want to control schema comparison on the global level.