Always derive collection types via schema
Currently, collection types (`Tuple` and `Array`) are a weird special
case, whereby collection type instances are normally not tracked by a
schema, except when used by another schema object.  This duality
requires a complex web of classes with lots of hacks to circumvent
typing violations.  The complexity of this implementation far outweighs
any possible benefit of deriving ephemeral collection types outside of the
schema.

This commit reworks the whole thing so that there are now only one class
for each collection kind with the usual traits of the schema object.
The changes are mostly about changing a bunch of functions that may
derive new collection types to return the modified schema object.