Fix ref propagation in certain cases of non-trivial inheritance
Consider the following (simplified) type structure:

      A
    /   \
    |    B
     \	/
      C

In this case, depending on the order of child processing, a ref created
on `A` might propagate to `C` *twice*, thus generating an "object
already exists" error.

To fix this, postpone the existence check until the action on the
descendant type is executed by leaning on `if_exists` and
`if_not_exists` knobs of the `AlterObject` and `CreateObject` commands.