Represent EdgeQL function bodies as Expressions
Currently, function bodies are always parsed and stored as string
literals, which are only compiled when the `CreateFunction` delta
command is applied.  This has obvious downsides: validation happens too
late, and there's no dependency tracing for objects referenced in the
function body, not to mention the ugly hacks in the parser that deparse
the `USING (<expr>)` form back to a string.

Co-authored-by: Paul Colomiets <paul@edgedb.com>