Always create a fenced scope node when compiling a statement
Currently, there's a cludge in `init_stmt()` that omits the
creation of a scope fence if `toplevel_stmt` is `None`.  The
reasoning is that the top-level statement would attach to the
root node of the scope tree.  However, in certain cases, such
as when a function is compiled, there may be multiple
"top-level" calls into the compiler with the same context, and
they would end up sharing the same scope node inappropriately.

There's no harm in an extra scope indirection at the root, so
just do it unconditionally.