Avoid running cardinality inference callbacks in temporary scopes
Since temporary scopes are thrown away upon exit from the context block,
any delayed cardinality inference scheduled while inside the block must
not proceed, because the scope is gone.  There is some protection from
this currently, but it's a porous patchwork, so replace it with a proper
solution and add explicit assertions that dangling scope pointers are
never referenced.