Treat defaults like implicit INSERT shape computables
Currently, in `INSERT` statements the defaults are treated like
schema-level computables, mostly for historical reasons, because schema
computables used to be defined via defaults.  That is not true anymore,
and the current formulation makes it hard to properly track the
cardinality of the default expression.  So, handle the defaults like
implicit shape computables instead.