Make sure module schemas are dropped in correct order
The underlying PostgreSQL schema must be dropped only when all other
objects have been.  The current coding tries to achieve that by placing
a low priority on the `DropSchema` operation, but fails to notice that
the operation is actually within a command group, so the priority should
be applied to the group instead.

Fixes: #1369.