Fix an issue with SDL ordering.
The computables were not tracing the dependency correctly, resulting in
occasional issues with SDL migrations.

When a computable contains a backwards link the dependency is now
using the correct link direction for tracking.

Also only pointers that have the same source as the computable are
"inherited" (to preserve link properties). All other paths get treated
as an arbitrary expression.

Fixes #1383