Improve performance of cached bootstrap
We currently cache generated stdlib artifacts, including the PL/pgSQL
script generated from the stdlib delta.  This saves time required to
generate the delta script, but not the cost of actually *running* the
script in Postgres, which has grown quite high, both in terms of CPU
time and memory requirements.  Fortunately, it's fairly easy to
efficiently cache that also by saving a dump of the template database.