Fix handling of functions with object arguments
Currently an attempt to reference any property of an object passed as an
argument to a function would crash the compiler.  The reason is that the
'source' aspect is advertised on the param rvar even though only the
identity of the object is actually passed to the function.  This is
actually a systemic problem, which I've been ad-hoc patching in similar
contexts before (see, e.g. c2b5e11edfe2).  The proper fix would be to
never advertise the 'source' aspect by default.

Fixes: #1335