Resolve dependency when an ALTER affects an existing expression.
When a schema change affects indexes, defaults, or function
bodies without changing their actual code, regenerate the corresponding
features, so that they are actually updated. We already know from
validation that the expression is still valid, so it's mainly a matter
of re-evaluating it in order to update type-related features.