Remove most direct references to schema objects at DDL>IR phase
The DDL AST -> Delta IR translation phase is not supposed to rely on
objects being present in the schema, and there are a few places where
that still happens, most notably in the handling of Aliases.  The latter
is because `CREATE ALIAS` currently tries to pretend it's really a
`CREATE TYPE ... USING (expr)`, and to know wheter it's `CREATE TYPE` or
`CREATE SCALAR TYPE` it needs to compile the alias expression at AST
translation phase.  Fix this by making aliases into thin wrappers around
the nested type command.  A side-effect of this is that we now have an
explicit `Alias` object in the schema.