Consistent handling of "DDL identifying" fields on schema objects
This rationalizes the construction of DDL IR for objects that require
extra information to identify them in addition to names, or those that
don't use names at all, like indexes.  Instead of a patchwork of attempt
to pass the necessary information in the form of `AlterObjectProperty`
subcommands, record it consistently as a field on the `ObjectCommand`
object alongside `classname`.