Normalize DDL and SDL expressions using fully-qualified names.
Normalize expressions used in DDL and SDL by enforcing the use of
fully-qualified names (based on the current module). This method leads
to more concise expressions as opposed to injecting a `WITH` block
setting a particular default module. This also makes it unnecessary to
keep the original expressions because the normalized form is very close
to the original form, so that indexes and constraints (as well as other
expressions such as computables) are still easily recognizable.

Fixes: #1254