Implement "script" output compilation mode
When a statement is executed as part of a multi-statement script, or
simply in the situation when the statement output is not needed, it
makes no sense to compile the query to produce `Data` messages to only
have them thrown away.

This adds a new `SCRIPT` output mode to the SQL compiler.  Currently, it
simply produces the cardinality of the result set, but can be extended
to return other statement-identifying information, which, for example,
might be useful to track error context.