Add DDL syntax to change whether a referenced schema object is "owned" (#1543)
Links, properties, constraints, indexes etc are either inherited from
the parent type(s), or are defined directly in a given type, in which
case the referenced object is "owned" by the type.  Currently, there is
no way to consistently express the change from "inherited" to "owned" or
vice-versa in DDL, so add the new `SET OWNED` and `DROP OWNED` clauses
to the appropriate `ALTER` commands.  This also renames the `is_local`
object field to `is_owned` everywhere for consistency.