Fix follow_imports configuration in mypy.ini
True and False are not actually valid values for follow_imports, but
mypy apparently only checks this when they are passed on the command line.
(Fix submitted to mypy: https://github.com/python/mypy/pull/9165).
This leads to unusual and inconsistent mypy behavior.

The intended values for follow_imports here are 'normal' and 'skip',
but it turns out that we do just fine if we set it to normal
everywhere, so do that.