Handle nested aliased types in GraphQL.
Aliased types have layout similar to their original tpyes, but with
potentially additional links or properties. This means that the GraphQL
reflected fields must have additional arguments on them. This, in turn,
means that the aliased type cannot be simply derived from the original
type's interface because the arguments would then have to match as well.
So aliased types are reflected without attempting to preserve inheritance,
but instead the focus is on making all the fields accessible.

Fixes #722.