Reimplement the pointer optionality qualifier check as a merge function
Current `REQUIRED` qualifier consistency check is done at an awkward
point in DDL AST visitor, which produces false positives when
`DROP REQUIRED` is actually a nop.  More importantly, though, the
current check does not prevent doing the wrong thing in the current
implementation of SDL.

Fix this by reimplementing the check as a field merge function, like
`cardinality` and `readonly`.