Optimize handling of singleton right operand in ??
When the cardinality of the right operand in ?? is known to be exactly
`ONE`, we can use a scalar SQL `COALESCE` without any precautions.  If
it's `AT_MOST_ONE`, use a fact that (SELECT (<empty_rel>)) is `NULL` to
avoid using `force_optional`, because the latter largely defeats the
optimization.

This also fixes cardinality inference of certain type intersection
paths.