Use ddl_identity for annotation refs in annotation value commands
Currently, reflection of annotation value commands relies on the
`annotation` reference to be present in the command, which is not always
the case, such as when only the annotation value is altered.  Use the
new DDL identity mechanism to always guarantee that the reference is
there in a canonical representation of a command.  While at it, do some
minor refactoring to make DDL identity handling sligtly more formal and
general.

Fixes: #1618.