Implement basic select-or-insert via coalesce
Implement a select-or-insert idiom using
`(SELECT ...) ?? (INSERT ...)`.

We only allow it under restricted circumstances (when the select and
insert both use the same value for an exclusive property) and it is
compiled using ON CONFLICT.

This doesn't work when the values are constants yet, because the
normalizer treats each constant separately. I'll change that in a
follow-up.

Mostly fixes #684, modulo follow-up.