Add an env variable to force using a cache even if the keys mismatch
This is useful for speeding up local testing in some cases and for
making it possible to run unit tests when there is an issue that
causes bootstrapping to fail (since failing unit tests are a lot
easier to narrow down).

I can rename the environment variable if somebody has a name they
prefer.