Fix broken errors in some correlated set cases
InvalidScopeConfiguration carries references to ScopeTreeNodes in it,
which causes grief because those fail to pickle due to the weakrefs in
them. (Though pickling them doesn't seem like the right thing to do
even if it wasn't for the weakref.)

Handle this in the same way that it is handled for the current cases
of this message that work: by catching it and reraising. (See
`register_set_in_scope`).