Fix bootstrap on RDS and make it faster overall
Bootstrap on RDS is currently broken because it attempts to restore the
bootstrap dump with incorrect credentials resulting in permission
errors.  Another issue is that the generated dump contains a `COMMENT ON
EXTENSION uuid-ossp` statement which fails because `COMMENT` is
unqualified and it finds the globally-configured extension, and, again,
fails with a permision error.

The fixes here:

1. Ensure that the extensions we depend on are created in a separate
   schema (`edgedbext`) and are commented out in the bootstrap dump.

2. Switch dump restoration from `psql` to `asyncpg.execute()`.  The
   latter is not only much faster due to there only being a single
   server roundtrip, it's also more robust, since `psql`, even with
   `--single-transaction` and `ON_ERROR_STOP`, does not do the right
   thing when the input is from stdin.

With this, bootstrap on a db.t3.micro instance takes 18 seconds.