Fix derived name mangling in schema
The current mangling schema isn't properly reversible, and we actually
rely on it being so.  Fix this by making sure that
`unmangle(mangle(X)) == X` always holds regardless of the depth of
mangling.

This creates an incompatibility with earlier dumps, because we carry an
internal `name -> id` map in them.  Fortunately, it's easy to fix by
retaining the old mangling code and using it to lookup in schema id maps
produced by older servers.