Fix broken errors in some correlated set cases (#1687)
InvalidScopeConfiguration carries references to ScopeTreeNodes in it,
which causes grief because those fail to pickle due to the weakrefs in
them. (Though pickling them doesn't seem like the right thing to do
even if it wasn't for the weakref.)

That infrastructure isn't actually used, so raise a more conventional
exception and thread in parser context information for useful spans.