Use "edgedb0" for system purposes and stop treating "edgedb" as special
The `edgedb` database is a database that we create by default, but it
should not be special: it can be dropped, and, accordingly, the server
should not rely on it being present.  Instead, use the `edgedb0`
template database for system purposes.  To do this, hoist the connection
block from PostgreSQL level to EdgeDB level.

Fixes: #1729