Overhaul how FOR interacts with iterators (#1735)
We now always compile FOR iterators when they are seen at SELECT
statements, rather than looking back up the tree to find them when we
encounter them at DML statements. At the compilation site, we check
whether there is DML in the body and hoist the iterator to a CTE if
so.

This fixes some issues involving nested iterators and mixed SELECT/DML
inside iterators and allows us to loosen the restrictions on
correlation of iterator variables.

It should make fixing #1699 (volatile keys with UNLESS CONFLICT ELSE)
practical. Hopefully it also clears the way for additional refactors.

Closes #1724. Closes #1674. Closes #1686.