Manage typeref and ptrref correspondence more carefully
Historical reliance on matching backend table column names for all
pointers in an inheritance hierarchy in turn led to sloppy use of parent
type pointer refs in subtype contexts.  The upcoming switch to
unique column names revealed all those places, so this commit adds more
diligence to maintaining the typeref/ptrref correpondence.  Things are
still a bit wobbly around handling of type intersection paths, but
that's not a new problem and will be addressed in a future patch.