Don't trigger recursion errors on medium-size set literals
Currently set literals get compiled into a chain of UNIONs, which
causes a big blowup in the nesting depth of the AST. As a result, we
trip a recursion error on set literals containing 60 one-element tuples,
which is a solidly medium-size input.

Fix this by compiling set literals into a tree of UNIONs instead of a
chain.

Fixes #475 to my satisfaction though others may disagree