Restore find_path_output as an optimization
This will fix an issue where fetch something twice for expr and
serialized, or similar, including sometimes doing two function calls!

A fix I made in #1762 made find_path_output no longer necessary for
correctness, so I removed it, but it will still important as an
optimization.

Restore the optimization bit of that code, and add a comment so I
don't forget and delete it again ;P.