Use pointer object ID instead of name for table column names
Currently, we use the name of a link or property as the name of a
corresponding column in the PostgreSQL table.  When we switched to the
use of object IDs for table names, column names were left as is, because
there was still reliance on table inheritance.  Now that table
inheritance was replaced with direct use of `UNION`, the requirement of
matching column names no longer exists and we can switch to
rename-resistant column mapping instead.