Introduce new enum definition syntax
Now to define an enum we use string literals in the type expression, as
in:

  CREATE SCALAR TYPE schema::Cardinality
    EXTENDING enum<'ONE', 'MANY'>;

With this commit, we switch to using identifiers:

  CREATE SCALAR TYPE schema::Cardinality
    EXTENDING enum<ONE, MANY>;

The motivation for the change is to nudge users towards defining enums
that can be easily reflected to other languages, such as TypeScript or
Python. The current syntax makes it easy to create enums with spaces
in them, evidenced by our own enums like schema::TargetDeleteAction.